Add run for >40%
